stable-pretraining/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "stable-pretraining"
description = "Stable Pretraining Library for Self-Supervised Learning"
authors = []
license = {text = "MIT"}
readme = {file = "README.md", content-type = "text/markdown"}
dependencies = [
    "torch",
    "torchvision",
    "torchmetrics",
    "lightning",
    "hydra-core",
    "omegaconf",
    "loguru",
    "tabulate",
    "pandas",
    "rich",
    "requests-cache",
    "prettytable",
    "submitit",
    "timm",
    "zstandard",
    "typer",
    "transformers",
    "pillow>=11.3.0",
    "wandb",
    "hydra-submitit-launcher",
    "matplotlib",
    "datasets",
    "pyarrow==20.0.0",
    "minari[hdf5]>=0.5.3",
    "scikit-learn",
    "richuru",
]

dynamic = ["version"]

[project.scripts]
spt = "stable_pretraining.cli:app"

[project.entry-points."lightning.pytorch.callbacks_factory"]
stablepretraining_callbacks = "stable_pretraining.callbacks.factories:default"

[tool.setuptools]
packages = ["stable_pretraining"]

[tool.setuptools_scm]
tag_regex = "^v(?P<version>.+)$"
write_to = "stable_pretraining/_version.py"
write_to_template = "version = \"{version}\"\n"

[project.optional-dependencies]
# Development and documentation tools
dev = [
    "pytest",
    "coverage",
    "pytest-cov",
    "codecov",
    "pre-commit",
    "ruff",
    "sphinx",
    "sphinx-gallery",
    "sphinx-book-theme",
    "sphinxcontrib-bibtex",
    "matplotlib",
    "memory_profiler",
    "myst-parser",
]


[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "D"]
ignore = ["D100", "D102", "D103", "D104", "D105", "D107"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true
